requirements.txt
pytz
Django==2.2.5
django-bootstrap4==1.0.1
django-debug-toolbar==2.0
django-extensions==2.2.1
django-filter==2.2.0
django-widget-tweaks==1.4.5
django-colorfield==0.1.15
python-decouple==3.1